+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -1,5 +1,10 @@
 ## v2.48.0-dev
 
+* Node v0.12.x users must run with --harmony. _This is the last release that
+    will support v0.12.x_
+* FIXED: the `webdriver.promise.ControlFlow` now has a consistent execution
+    order for tasks/callbacks scheduled in different turns of the JS event loop.
+    Refer to the `webdriver.promise` documentation for more details.
 * FIXED: a single `firefox.Binary` instance may be used to configure and
     launch multiple FirefoxDriver sessions.